Suggestion manager and logs
Added a suggestion manager to improve reviewing suggestions as well as a basic log system. Channel and tag IDs are using IDs from my test server, so they will need to be changed when pushed to the main bot. Will prob add an easier way of defining the IDs without needing to edit the code at some point.